@@ -4,17 +4,14 @@
     "keywords": ["log", "logging", "logger", "handler"],
     "type": "library",
     "license": "BSD-3-Clause",
-    "homepage": "http://php5.laurent-laville.org/callbackfilterhandler/",
     "support": {
         "source": "https://github.com/llaville/monolog-callbackfilterhandler",
         "issues": "https://github.com/llaville/monolog-callbackfilterhandler/issues"
     },
     "require": {
-        "php": "^5.3|^7.0",
-        "monolog/monolog": "^1.10"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^6.0|^7.0|^8.0|^9.0"
+        "php": "^7.2|^8.0",
+        "ext-json": "*",
+        "monolog/monolog": "^2.0"
     },
     "authors": [
         {
@@ -26,15 +23,12 @@
     ],
     "autoload": {
         "psr-4": {
-            "Bartlett\\": "src/Bartlett"
+            "Bartlett\\Monolog\\Handler\\": "src/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Bartlett\\Tests\\Monolog\\Handler\\": "tests/"
+            "Bartlett\\Monolog\\Handler\\Tests\\": "tests/"
         }
-    },
-    "scripts": {
-        "tests": "vendor/bin/phpunit"
     }
 }
