@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /**
  * Callback Filter Handler for Monolog.
  *
@@ -7,14 +7,20 @@
  * @author   Laurent Laville <pear@laurent-laville.org>
  * @author   Christophe Coevoet
  * @license  http://www.opensource.org/licenses/bsd-license.php  BSD License
- * @version  GIT: $Id$
- * @link     http://php5.laurent-laville.org/callbackfilterhandler/
  */
 
 namespace Bartlett\Monolog\Handler;
 
 use Monolog\Handler\AbstractHandler;
 use Monolog\Handler\HandlerInterface;
+use Monolog\Handler\ProcessableHandlerInterface;
+use Monolog\Logger;
+
+use RuntimeException;
+use function array_shift;
+use function array_unshift;
+use function is_callable;
+use function json_encode;
 
 /**
  * Monolog handler wrapper that filters records based on a list of callback functions.
@@ -24,17 +30,21 @@ use Monolog\Handler\HandlerInterface;
  * @author   Laurent Laville <pear@laurent-laville.org>
  * @author   Christophe Coevoet
  * @license  http://www.opensource.org/licenses/bsd-license.php BSD License
- * @version  Release: @package_version@
  * @since    Class available since Release 1.0.0
  */
-class CallbackFilterHandler extends AbstractHandler
+class CallbackFilterHandler extends AbstractHandler implements ProcessableHandlerInterface
 {
     /**
      * Handler or factory callable($record, $this)
      *
-     * @var callable|\Monolog\Handler\HandlerInterface
+     * @var callable|HandlerInterface
      */
     protected $handler;
+
+    /**
+     * @var int
+     */
+    protected $handlerLevel;
 
     /**
      * Filters callable to restrict log records.
@@ -44,35 +54,39 @@ class CallbackFilterHandler extends AbstractHandler
     protected $filters;
 
     /**
-     * Whether the messages that are handled can bubble up the stack or not
+     * Changes to apply on log records, by a stack of callable
      *
-     * @var boolean
+     * @var callable[]
      */
-    protected $bubble;
+    protected $processors;
 
     /**
      * @param callable|HandlerInterface $handler Handler or factory callable($record, $this).
      * @param callable[]                $filters A list of filters to apply
+     * @param int|string                $level   The minimum logging level at which this handler will be triggered
      * @param boolean                   $bubble  Whether the messages that are handled can bubble up the stack or not
      */
-    public function __construct($handler, array $filters, $bubble = true)
+    public function __construct($handler, array $filters, $level= Logger::DEBUG, bool $bubble = true)
     {
-        $this->handler = $handler;
-        $this->bubble  = $bubble;
-        $this->filters = array();
-
-        if (!$this->handler instanceof HandlerInterface) {
-            if (!is_callable($this->handler)) {
-                throw new \RuntimeException(
-                    "The given handler (" . json_encode($this->handler)
+        if (!$handler instanceof HandlerInterface) {
+            if (!is_callable($handler)) {
+                throw new RuntimeException(
+                    "The given handler (" . json_encode($handler)
                     . ") is not a callable nor a Monolog\\Handler\\HandlerInterface object"
                 );
             }
         }
 
+        $this->handlerLevel = $level;
+        parent::__construct($level, $bubble);
+
+        $this->handler = $handler;
+        $this->filters = [];
+        $this->processors = [];
+
         foreach ($filters as $filter) {
             if (!is_callable($filter)) {
-                throw new \RuntimeException(
+                throw new RuntimeException(
                     "The given filter (" . json_encode($filter)
                     . ") is not a callable object"
                 );
@@ -82,18 +96,35 @@ class CallbackFilterHandler extends AbstractHandler
     }
 
     /**
+     * {@inheritDoc}
+     */
+    public function pushProcessor(callable $callback): HandlerInterface
+    {
+        array_unshift($this->processors, $callback);
+        return $this;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function popProcessor(): callable
+    {
+        return array_shift($this->processors);
+    }
+
+    /**
      * {@inheritdoc}
      */
-    public function isHandling(array $record)
+    public function isHandling(array $record): bool
     {
-        if ($record['level'] < $this->handler->getLevel()) {
+        if ($record['level'] < $this->handlerLevel) {
             return false;
         }
 
-        if (array_key_exists('message', $record)) {
+        if (isset($record['message'])) {
             // when record is full filled, try each filter
             foreach ($this->filters as $filter) {
-                if (!call_user_func($filter, $record, $this->handler->getLevel())) {
+                if (!$filter($record, $this->handlerLevel)) {
                     return false;
                 }
             }
@@ -104,7 +135,7 @@ class CallbackFilterHandler extends AbstractHandler
     /**
      * {@inheritdoc}
      */
-    public function handle(array $record)
+    public function handle(array $record): bool
     {
         if (!$this->isHandling($record)) {
             return false;
@@ -112,15 +143,15 @@ class CallbackFilterHandler extends AbstractHandler
 
         // The same logic as in FingersCrossedHandler
         if (!$this->handler instanceof HandlerInterface) {
-            $this->handler = call_user_func($this->handler, $record, $this);
+            $this->handler = ($this->handler)($record, $this);
             if (!$this->handler instanceof HandlerInterface) {
-                throw new \RuntimeException("The factory callable should return a HandlerInterface");
+                throw new RuntimeException("The factory callable should return a HandlerInterface");
             }
         }
 
         if ($this->processors) {
             foreach ($this->processors as $processor) {
-                $record = call_user_func($processor, $record);
+                $record = ($processor)($record);
             }
         }
 
@@ -132,9 +163,9 @@ class CallbackFilterHandler extends AbstractHandler
     /**
      * {@inheritdoc}
      */
-    public function handleBatch(array $records)
+    public function handleBatch(array $records): void
     {
-        $filtered = array();
+        $filtered = [];
         foreach ($records as $record) {
             if ($this->isHandling($record)) {
                 $filtered[] = $record;
